fix : expense amount 타입

+++ src/app/(routing)/trip/[tid]/_components/CreateExpenseModal.tsx
@@ -194,7 +194,7 @@ const ExpenseInputContent = () => {
             <span className="text-red-300">
               {(
                 (totalIncomes || 0) -
-                (totalExpenses + parseInt(expenseData.amount) || 0)
+                (totalExpenses + parseFloat(expenseData.amount) || 0)
               ).toLocaleString()}
             </span>
             <span className="text-red-300 text-base">/</span>
